# cmd/api/docs/swagger.yaml
definitions:
...
  req.Category:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  req.Product:
    properties:
      brand:
        type: string
      categoryid:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - brand
    - categoryid
    - description
    - name
    type: object
  req.ProductItems:
    properties:
      color:
        type: string
      gender:
        type: string
      material:
        type: string
      model:
        type: string
      price:
        type: number
      productId:
        type: integer
      qty:
        type: integer
      size:
        type: integer
      sku:
        type: string
    type: object
...
  res.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/admin/category/add:
    post:
      consumes:
      - application/json
      description: Admin can create new category from admin panel
      operationId: create-category
      parameters:
      - description: New category name
        in: body
        name: category_name
        required: true
        schema:
          $ref: '#/definitions/req.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Create new product category
      tags:
      - Category
  /api/admin/category/delete/{category_id}:
    delete:
      consumes:
      - application/json
      description: Admin can delete a category
      operationId: delete-category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin can delete a category
      tags:
      - Category
  /api/admin/category/listall:
    get:
      consumes:
      - application/json
      description: Admin, users and unregistered users can see all the available categories
      operationId: view-all-categories
      parameters:
      - description: Page number for pagination
        in: query
        name: count
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: View all available categories
      tags:
      - Category
  /api/admin/category/update/{id}:
    patch:
      consumes:
      - application/json
      description: Admin can update category details
      operationId: update-category
      parameters:
      - description: ID of the Category to be updated
        in: path
        name: id
        required: true
        type: string
      - description: category info
        in: body
        name: category_details
        required: true
        schema:
          $ref: '#/definitions/req.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin can update category details
      tags:
      - Category
...
  /api/admin/product-item/add:
    post:
      consumes:
      - application/json
      description: add a new product item with an existing product id
      operationId: add-product-item
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/req.ProductItems'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new product item
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Failed to add new product item
          schema:
            $ref: '#/definitions/res.Response'
      summary: Add products item
      tags:
      - Product Item
  /api/admin/product-item/update:
    patch:
      consumes:
      - application/json
      description: update existing product item with id
      operationId: update-product-item
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/req.ProductItems'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new product item
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Failed to add new product item
          schema:
            $ref: '#/definitions/res.Response'
      summary: Add products item
      tags:
      - Product Item
  /api/admin/product/{id}:
    get:
      consumes:
      - application/json
      description: list all saved products
      operationId: display-product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new product item
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Failed to add new product item
          schema:
            $ref: '#/definitions/res.Response'
      summary: Display products
      tags:
      - Product
  /api/admin/product/add:
    post:
      consumes:
      - application/json
      description: This endpoint allows an admin user to create a new product item.
      operationId: add-product
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/req.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new product item
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Failed to add new product item
          schema:
            $ref: '#/definitions/res.Response'
      summary: Creates a new produc
      tags:
      - Product
  /api/admin/product/update/{id}:
    patch:
      consumes:
      - application/json
      description: updating exsisting product details.
      operationId: update-product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/req.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated new product
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Failed to add new product
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update product
      tags:
      - Product
...
  /api/user/category/listspecific/{category_id}:
    get:
      consumes:
      - application/json
      description: Users and admins can fetch details of a specific category using
        id
      operationId: display-category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Fetch details of a specific category using category id
      tags:
      - Category
...
  /api/user/product-item/{product_id}:
    get:
      consumes:
      - application/json
      description: update existing product item with id
      operationId: display-product-items
      parameters:
      - description: id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new product item
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Failed to add new product item
          schema:
            $ref: '#/definitions/res.Response'
      summary: display products item
      tags:
      - Product Item
...
swagger: "2.0"